Add find consensus task & update tests
